split grid fragment out, for use with other types
